Implement episodic learning. ---------------------------- Every episode includes multiple epochs, and at the end of each episode the current controller is used to rollout the system. At the start of training, the nominal controller is used for the rollout instead. This required a major refactor of the datamodules and dynamical systems, so now the safe/unsafe/goal regions are contained in the dynamical systems instead of within the datamodules.
Implement primal/dual optimization
----------------------------------
Previously, we relied on hand-tuned weights for the constraint losses relative to the objective losses. Now we make that difference explicit, and each
constraint is assigned a Lagrange multiplier, and the Lagrange multipliers are tuned using stochastic gradient ascent with weight decay, allowing us to
fine-tune these weights as the training proceeds.