Pretrained contraction metric (#3)
Add an experimental workflow that enables pre-training of the contraction metric on trajectory data. 

The potential benefits include not needing to jointly learn a controller to define a contraction metric. 
However, I am unsure if the training objective defined in `TwoStageTrainer.pretrain_contraction_loss_M` is correct. 
As for empirical results: we do see some evidence that the pre-training increases sample-efficiency of learning the policy.  However, it also sometimes impairs learning of a robust tracking policy, so the jury's out on whether this makes sense